% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosp_ensemble_empirical_data_functions.R
\name{construct_hosp_url}
\alias{construct_hosp_url}
\title{Construct hosp url}
\usage{
construct_hosp_url(source = c("state", "facility"))
}
\arguments{
\item{source}{string either "state" or "facility"; used internally by the function to pull the fixed
identifier for the dataset. Only "state" is currently implemented.}
}
\value{
string url to data; this can be fed to \code{readr::read_csv}, \code{datatable::fread()} or other similar function
}
\description{
Construct a time series url for pulling the latest time series data
}
\examples{
construct_hosp_url("state")
}
