% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_functions.R
\name{coverage_predictions}
\alias{coverage_predictions}
\title{Get coverage of predictions}
\usage{
coverage_predictions(pred_min, pred_max, empir_obs, inclusive = T)
}
\arguments{
\item{pred_min}{vector with predicted minimum}

\item{pred_max}{vector with predicted maximum}

\item{empir_obs}{vector with observed outcomes}

\item{inclusive}{logical, default = T, coverage is estimated using inclusive bounds, rather than exclusive}
}
\value{
count (numerator), possible (denominator), and proportion of times (numerator/denominator) that empir_obs falls within pred_min and pred_max, inclusive
}
\description{
Function allows for estimating of coverage of xmin,xmax, given empirical y
}
\examples{
coverage_predictions(lower, upper, observed)
coverage_predictions(lower, upper, observed, inclusive=F)
}
