% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosp_ensemble_data_processing_functions.R
\name{estimates_consistent}
\alias{estimates_consistent}
\title{Check if estimates vary by location and target_end date}
\usage{
estimates_consistent(hosp, locvar = "location", datevar = "target_end_date")
}
\arguments{
\item{hosp}{The raw hospitalization data}

\item{locvar}{string name of location variable (default = "location")}

\item{datevar}{string name of date variable (default = "target_end_date"); could use "target"}
}
\value{
boolean indicator; if TRUE if the dates are consistent, else false
}
\description{
Function will return boolean indicator as to whether or not a set of hospitalization data
by group (i.e. quantiles functions) varying in number over location and date
}
\examples{
estimates_consistent(hosp)
}
