% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosp_ensemble_plotting_functions.R
\name{forecast_plot}
\alias{forecast_plot}
\title{Generate a labeled ensemble plot, with individual model estimates}
\usage{
forecast_plot(
  ensemble,
  indivmodels,
  location_var = "location",
  location = NULL,
  type = NULL,
  intervals = c(50, 95)
)
}
\arguments{
\item{ensemble}{This is the ensemble data (i.e. produced by generate_slp_ensemble, for example)}

\item{location_var}{String variable that indicates the name of the variable that indicates location (default is "location")}

\item{location}{string variable indicating the location that is being plotting (for plot title purposes only)}

\item{type}{string variable that describes what type of ensemble is being plotted}

\item{intervals}{vector of two interval sizes (default 50 and 95)}

\item{indivmodesl}{This is the data table containing the hospitalization data (i.e. produced by hosp_processing)}
}
\value{
returns a plot object
}
\description{
Function will receive a data frame with both individual models and the ensemble
and return a plot with some labels
}
\examples{
forecast_plot(ensdata, hospdata,location="US",type="SLP")
forecast_plot(ensdata, hospdata)
}
