% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_functions.R
\name{get_coverage_stats}
\alias{get_coverage_stats}
\title{Get Coverage Stats from a Merged Data Frame}
\usage{
get_coverage_stats(dat, bycols = NULL, qnames = NULL)
}
\arguments{
\item{dat}{This is the data frame of merged hospitalization predictions and empirical data}

\item{bycols}{(default is NULL) a string vector of columns to group by}

\item{qnames}{(default is NULL) a string vector of the quantile column names; when NULL, the function will look for
23 columns in \code{dat} starting with 0; if not found will return error}
}
\value{
data table of coverage predictions
}
\description{
This function takes a merged data frame, and calculates coverage stats by any column; the format
of the merged data frame is important. It must have both hospital and predicted values. The
predicted values must be by quantile, with names like \code{0.5}, etc.. The function will look for
these names, and expects to find them
}
\examples{
get_coverage_stats(dat=dat)
get_coverage_stats(dat=dat,bycols=c("horizon","indicator"))
}
