% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosp_ensemble_data_processing_functions.R
\name{get_data_for_model}
\alias{get_data_for_model}
\title{Get raw data for a modeling group}
\usage{
get_data_for_model(
  modelname,
  date,
  context = "../covid19-forecast-hub/data-processed/"
)
}
\arguments{
\item{modelname}{character name of the model (must be subfolder in context)}

\item{date}{yyyy-mm-dd character indicating the date threshold for the model}

\item{context}{character path to the data-processed folder for covid19-forecast-hub}
}
\value{
data table containing the data for this modeling group
}
\description{
A function to get the last dataset, dated up to a given date, for a given modeling group. The function
will attempt to find the single matching csv file for this modeling group that is up-to, but
not exceeding the date provided
}
\examples{
get_data_for_model(modelname="JHU-APL",date="2020-11-01")
get_data_for_model(modelname="JHU-APL",date="2020-11-01",context =mypath_to_covid_hub)
}
