% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosp_ensemble_data_processing_functions.R
\name{get_hosp_raw_data}
\alias{get_hosp_raw_data}
\title{Get the hospitalization quantile functions by group}
\usage{
get_hosp_raw_data(
  model_dir_list,
  date_threshold,
  date_limits = c(as.Date(date_threshold), as.Date(date_threshold) + 28),
  context = "../covid19-forecast-hub/data-processed/",
  includepoint = F
)
}
\arguments{
\item{model_dir_list}{a list of character strings, one for each subfolder in the data-processed folder for covid19-forecast-hub}

\item{date_threshold}{a date threshold - only data prior to this date will be pulled; this data will also be used to restrict
the forecast date for any data sets returned. Only if the forecast date is greater than or equal to
six days prior to the forecast date, will the data be retained.}

\item{date_limits}{a two Date vector of limits on the data to be returned. The default is to return data from the threshold date
to 28 days in the future}

\item{context}{character path to the data-processed folder for covid19-forecast-hub}

\item{includepoint}{default FALSE (normally, we will drop the point estimate)}
}
\value{
raw hospitalization quantile functions by modeling group
}
\description{
This function is the workhorse of the pre-processing set of functions. Given a list of modeling
group subfolders, and a date threshold, it will pull the raw data (if available),
limit the data to hospitalization outcomes, and limit the dates to certain parameters
}
\examples{
get_hosp_raw_data(mydirs, "2020-11-02")
}
