% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosp_ensemble_data_processing_functions.R
\name{get_model_directories}
\alias{get_model_directories}
\title{Return a list of model directories}
\usage{
get_model_directories(context = "../covid19-forecast-hub/data-processed/")
}
\arguments{
\item{context}{character path to the data-processed folder for covid19-forecast-hub}
}
\value{
list of subfolders in the covid19 process data subfolder
}
\description{
Helper function to get the list of model groups in the covid19-hub processed data collection
}
\examples{
get_model_directories()
get_model_directories(context =mypath_to_covid_hub)
}
