% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosp_ensemble_data_processing_functions.R
\name{hosp_processing}
\alias{hosp_processing}
\title{Check consistency of hospitalization data}
\usage{
hosp_processing(
  hosp,
  datethreshold,
  valid_models = valid_hospitalization_models(),
  cu_select_only = TRUE,
  use_target_label_28 = FALSE
)
}
\arguments{
\item{hosp}{Data table containing the modeling groups raw quantile functions by location and date}

\item{datethreshold}{string date "YYYY-MM-DD" indicating the forecast date}

\item{valid_models}{a list of valid hospitalization modeling groups. The default is to pull from
\code{valid_hospitalization_models()} which is a hard coded list available in the package}

\item{cu_select_only}{logical default = T include only the CU-select model if there are multiple Columbia Models}

\item{use_target_label_28}{logical default = F, use the target label to isolate 28 days of forecast}
}
\value{
TRUE if all tests are passed; otherwise error condition
}
\description{
Function provides some checks on the hospitalization data
}
\examples{
hosp_processing(hosp)
}
