% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosp_ensemble_plotting_functions.R
\name{make_model_caption}
\alias{make_model_caption}
\title{Helper function to create contributing models caption, with wrapping}
\usage{
make_model_caption(modelnames, perline = 6)
}
\arguments{
\item{perline}{integer value of when to make line breaks, i.e. every \code{perline}; default
is 6}

\item{modelname}{character vector of modelnames}
}
\value{
a character caption
}
\description{
Function takes a list of modelnames and returns a caption, with
line breaks very \code{perline} components
}
\examples{
make_model_caption(mymodels,4)
}
\keyword{internal}
