% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosp_ensemble_plotting_functions.R
\name{plots_by_location}
\alias{plots_by_location}
\title{Generate all plots given an ensemble, and hospitalization frames}
\usage{
plots_by_location(
  ensemble,
  indivmodels,
  location_var,
  type = NULL,
  intervals = c(50, 95)
)
}
\arguments{
\item{ensemble}{This is the ensemble data (i.e. produced by generate_slp_ensemble, for example)}

\item{indivmodels}{This is the data table containing the hospitalization data (i.e. produced by hosp_processing)}

\item{location_var}{String variable name found in both ensemble and indivmodels that splits by location, and will be used as the title}

\item{type}{string variable that describes what type of ensemble is being plotted (for labeling only, default is NULL)}

\item{intervals}{vector of two interval sizes (default 50 and 95)}
}
\value{
returns a frame of plots
}
\description{
Function will receive a data frame with both individual models and the ensemble
and, for each location in the combined frames, will call the forecast plot function.
This is simply a convenience function around the above
}
\examples{
forecast_plot(data,"US","SLP")
}
