% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosp_ensemble_empirical_data_functions.R
\name{pull_and_process_state}
\alias{pull_and_process_state}
\title{Function to pull the and pre-process the raw state level file}
\usage{
pull_and_process_state(url)
}
\arguments{
\item{url}{The url for the data}
}
\description{
Function will pull and process the raw state level file.
}
\keyword{internal}
