% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosp_ensemble_data_processing_functions.R
\name{show_inconsistent}
\alias{show_inconsistent}
\title{Show locations where the number of estimates vary by location and target_end date}
\usage{
show_inconsistent(
  hosp,
  locvar = "location",
  datevar = "target_end_date",
  locations = NULL
)
}
\arguments{
\item{hosp}{The raw hospitalization data}

\item{locvar}{string name of location variable (default = "location")}

\item{datevar}{string name of date variable (default = "target_end_date"); could use "target"}

\item{locations}{vector of character location values (default is NULL); if this is provided, the function
will return a list of data tables; one for each location, and each one containing the list
of the modeling groups, and the start and end date (and number of days) of each group}
}
\value{
list of locations (if \code{locations} is NULL) or a list of data tables showing the inconsistencies; one
data table per value in \code{locations}
}
\description{
Function will return a list of locations, along with the number of estimates at each
date, if the number of estimates varies over the date range
}
\examples{
show_inconsistent(hosp)
}
