% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosp_ensemble_plotting_functions.R
\name{state_plot}
\alias{state_plot}
\title{Generate a plot of individual modeling group raw data}
\usage{
state_plot(
  hosp,
  target_var = "value",
  exclude_models = NULL,
  plt_title = NULL,
  legendpos = c("bottom", "right", "left", "top"),
  legrows = NULL
)
}
\arguments{
\item{hosp}{data frame of hospitalization data; preferably filtered down to a single location}

\item{target_var}{string name of the hospitalization field (default is "value")}

\item{exclude_models}{optional string vector of names of source models to exclude}

\item{plt_title}{option plot title}

\item{legendpos}{optional legend position must be one of "bottom" (default),"right", "left", or "top"}

\item{legrows}{optional number of legend rows for models (default will be determined by ggplot)}
}
\value{
plot of the individual modeling group's median estimates of the forecasted hospitalizations
}
\description{
This function will return a plot of the modeling group specific median estimates; the inputted
data frame should only recieve a single location; model groups can be optionally excluded
}
\examples{
state_plot(hosp[location==24], target_var="hospitalizations")
state_plot(hosp[location==24], target_var="value", exclude_models=c("CU-scenario_high"), plt_title("Maryland"))
}
