% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosp_ensemble_data_processing_functions.R
\name{valid_hospitalization_models}
\alias{valid_hospitalization_models}
\title{Get a current list of valid hospitalization models}
\usage{
valid_hospitalization_models()
}
\value{
list of valid hospitalization models
}
\description{
Function will return a list of current valid hospitalization models
}
\examples{
valid_hospitalization_models()
}
