% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosp_ensemble_empirical_data_functions.R
\name{pull_empirical_hospitalization_data}
\alias{pull_empirical_hospitalization_data}
\title{Pull empirical hospitalization data}
\usage{
pull_empirical_hospitalization_data(source = c("state"))
}
\arguments{
\item{source}{string either "state" or "facility"; used internally by the function to pull the fixed
identifier for the dataset. If "facility", the dataset at
https://healthdata.gov/dataset/covid-19-reported-patient-impact-and-hospital-capacity-facility will be pulled.
If "state" (default), the dataset at
https://healthdata.gov/dataset/covid-19-reported-patient-impact-and-hospital-capacity-state-timeseries will
be returned. Only state is currently implemented.}
}
\value{
data.table of hospitalization data
}
\description{
As of 2020/11/16, healthdata.gov began providing publicly available
csv formatted data of hospitalizations associated with covid. This
function takes a url, reads in the timeseries data, and converts it
to long format,combining pediatric and adult, and creating a US version
}
\examples{
pull_empirical_hospitalization_data()
pull_empirical_hospitalization_data(url)
pull_empirical_hospitalization_data(url=hosp_url)
}
